requirements.txt
#pandas-datareader = 0.8.1
#googlefinance = 0.7
#yahoo-finance = 1.4.0
python-dotenv = 0.12.0
psycopg2
pandas
python-box